+++ src/main/java/com/safaria/backend/repository/TourScheduleRepository.java
@@ -4,8 +4,8 @@ import com.safaria.backend.entity.TourSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface TourScheduleRepository extends JpaRepository<TourSchedule, Integer> {
+public interface TourScheduleRepository extends JpaRepository<TourSchedule, Long> {
 
-    List<TourSchedule> findByTour_TourId(Integer tourId);
+    List<TourSchedule> findByTour_TourId(Long tourId);
 
 }
